Extract DocUpdateMessages generation to activity-server plugin
