+++ dev/tool/package.json
@@ -107,6 +107,8 @@
     "@hcengineering/server-tracker-resources": "^0.6.0",
     "@hcengineering/server-view": "^0.6.0",
     "@hcengineering/server-view-resources": "^0.6.0",
+    "@hcengineering/server-activity": "^0.6.0",
+    "@hcengineering/server-activity-resources": "^0.6.0",
     "@hcengineering/setting": "^0.6.11",
     "@hcengineering/tags": "^0.6.12",
     "@hcengineering/task": "^0.6.13",
