+++ models/all/package.json
@@ -77,6 +77,7 @@
     "@hcengineering/model-request": "^0.6.0",
     "@hcengineering/model-server-request": "^0.6.0",
     "@hcengineering/model-server-view": "^0.6.0",
+    "@hcengineering/model-server-activity": "^0.6.0",
     "@hcengineering/model-server-openai": "^0.6.0",
     "@hcengineering/model-server-translate": "^0.6.0",
     "@hcengineering/model-support": "^0.6.0"
