+++ models/all/src/migration.ts
@@ -36,8 +36,8 @@ import { boardOperation } from '@hcengineering/model-board'
 import { hrOperation } from '@hcengineering/model-hr'
 import { bitrixOperation } from '@hcengineering/model-bitrix'
 import { calendarOperation } from '@hcengineering/model-calendar'
-import { notificationServerOperation } from '@hcengineering/model-server-notification'
 import { activityOperation } from '@hcengineering/model-activity'
+import { activityServerOperation } from '@hcengineering/model-server-activity'
 
 export const migrateOperations: [string, MigrateOperation][] = [
   ['core', coreOperation],
@@ -62,5 +62,5 @@ export const migrateOperations: [string, MigrateOperation][] = [
   ['hr', hrOperation],
   ['bitrix', bitrixOperation],
   ['inventiry', inventoryOperation],
-  ['notificationServer', notificationServerOperation]
+  ['activityServer', activityServerOperation]
 ]
