+++ models/server-activity/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@hcengineering/model-server-notification",
+  "name": "@hcengineering/model-server-activity",
   "version": "0.6.0",
   "main": "lib/index.js",
   "author": "Anticrm Platform Contributors",
@@ -15,12 +15,12 @@
   "devDependencies": {
     "@hcengineering/platform-rig": "^0.6.0",
     "@typescript-eslint/eslint-plugin": "^6.11.0",
-    "@typescript-eslint/parser": "^6.11.0",
-    "eslint": "^8.54.0",
-    "eslint-config-standard-with-typescript": "^40.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.4.0",
     "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-n": "^15.4.0",
+    "eslint": "^8.54.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "eslint-config-standard-with-typescript": "^40.0.0",
     "prettier": "^3.1.0",
     "prettier-plugin-svelte": "^3.1.0"
   },
@@ -28,12 +28,9 @@
     "@hcengineering/activity": "^0.6.0",
     "@hcengineering/core": "^0.6.28",
     "@hcengineering/model": "^0.6.7",
-    "@hcengineering/model-chunter": "^0.6.0",
-    "@hcengineering/model-core": "^0.6.0",
-    "@hcengineering/model-notification": "^0.6.0",
-    "@hcengineering/notification": "^0.6.16",
     "@hcengineering/platform": "^0.6.9",
-    "@hcengineering/server-core": "^0.6.1",
-    "@hcengineering/server-notification": "^0.6.1"
+    "@hcengineering/server-activity": "^0.6.0",
+    "@hcengineering/server-activity-resources": "^0.6.0",
+    "@hcengineering/server-core": "^0.6.1"
   }
 }
