+++ server-plugins/activity/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@hcengineering/model-server-notification",
+  "name": "@hcengineering/server-activity",
   "version": "0.6.0",
   "main": "lib/index.js",
   "author": "Anticrm Platform Contributors",
@@ -9,29 +9,28 @@
     "build:watch": "tsc",
     "lint:fix": "eslint --fix src",
     "lint": "eslint src",
-    "format": "format src"
+    "format": "format src",
+    "test": "jest --passWithNoTests --silent"
   },
-  "template": "@hcengineering/model-package",
   "devDependencies": {
     "@hcengineering/platform-rig": "^0.6.0",
+    "@types/node": "~16.11.12",
     "@typescript-eslint/eslint-plugin": "^6.11.0",
-    "@typescript-eslint/parser": "^6.11.0",
-    "eslint": "^8.54.0",
-    "eslint-config-standard-with-typescript": "^40.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.4.0",
     "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-n": "^15.4.0",
+    "eslint": "^8.54.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "eslint-config-standard-with-typescript": "^40.0.0",
     "prettier": "^3.1.0",
+    "typescript": "^5.2.2",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "@types/jest": "^29.5.5",
     "prettier-plugin-svelte": "^3.1.0"
   },
   "dependencies": {
-    "@hcengineering/activity": "^0.6.0",
     "@hcengineering/core": "^0.6.28",
-    "@hcengineering/model": "^0.6.7",
-    "@hcengineering/model-chunter": "^0.6.0",
-    "@hcengineering/model-core": "^0.6.0",
-    "@hcengineering/model-notification": "^0.6.0",
-    "@hcengineering/notification": "^0.6.16",
     "@hcengineering/platform": "^0.6.9",
     "@hcengineering/server-core": "^0.6.1",
     "@hcengineering/server-notification": "^0.6.1"
