+++ server-plugins/activity/src/index.ts
@@ -1,0 +1,36 @@
+//
+// Copyright © 2022 Hardcore Engineering Inc.
+//
+// Licensed under the Eclipse Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License. You may
+// obtain a copy of the License at https://www.eclipse.org/legal/epl-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import { Plugin, Resource } from '@hcengineering/platform'
+import { plugin } from '@hcengineering/platform'
+import type { TriggerFunc } from '@hcengineering/server-core'
+
+export * from './types'
+export  * from './utils'
+
+/**
+ * @public
+ */
+export const serverActivityId = 'server-activity' as Plugin
+
+
+/**
+ * @public
+ */
+export default plugin(serverActivityId, {
+  trigger: {
+    ActivityMessagesHandler: '' as Resource<TriggerFunc>
+  },
+})
