+++ server-plugins/notification-resources/src/types.ts
@@ -1,5 +1,3 @@
-import { Hierarchy, ModelDb, Storage, TxFactory } from '@hcengineering/core'
-
 /**
  * @public
  */
@@ -7,11 +5,4 @@ export interface Content {
   text: string
   html: string
   subject: string
-}
-
-export interface NotificationControl {
-  findAll: Storage['findAll']
-  hierarchy: Hierarchy
-  txFactory: TxFactory
-  modelDb: ModelDb
 }
